requirements.txt
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
PyYAML>=6.0.1
httpx>=0.27.0
tenacity>=8.2.3
APScheduler>=3.10.4
